EY-2646: Litt separation of concerns. Vi gjer så mykje greier med trygdetid her at det er fint å få det ut i eiga service
